update get ready copy
There is an automated email sent by the CE Cart which includes
official-sounding username and password information and invites them to
setup their MCAD.edu account. Their MCAD.edu account is NOT their
official MCAD account, and so students are confused.

The webmaster is working on eliminating these confusing emails.

This includes orientation copy edits, which hopefully reduces the number
of support calls and clarifies how to get started.